refactor: mainPage class size optimization, add dualSlider class
